Unchanged tests for docs as well.
