+++ test/saved-responses/docs/filter only.json
@@ -1,0 +1,48 @@
+{
+  "summary": {
+    "searchParam": {
+      "filter": "pid:PBsve435",
+      "indexname": "test",
+      "sort": "field:pid",
+      "waitfortotal": "true",
+      "wordsaroundhit": "1"
+    },
+    "searchTime": "VALUE_REMOVED",
+    "countTime": "VALUE_REMOVED",
+    "windowFirstResult": 0,
+    "requestedWindowSize": 50,
+    "actualWindowSize": 1,
+    "windowHasPrevious": false,
+    "windowHasNext": false,
+    "stillCounting": false,
+    "numberOfDocs": 1,
+    "numberOfDocsRetrieved": 1,
+    "docFields": {
+      "pidField": "pid",
+      "titleField": "title"
+    },
+    "metadataFieldDisplayNames": {
+      "fromInputFile": "From input file",
+      "pid": "Pid",
+      "title": "Title"
+    }
+  },
+  "docs": [
+    {
+      "docPid": "PBsve435",
+      "docInfo": {
+        "fromInputFile": [
+          "/input/PBsve435.xml"
+        ],
+        "pid": [
+          "PBsve435"
+        ],
+        "title": [
+          "service encounter about getting an address printed out"
+        ],
+        "lengthInTokens": 164,
+        "mayView": false
+      }
+    }
+  ]
+}
