--- test/test/hits-unchanged.js
@@ -1,77 +1,0 @@
-const chai = require("chai");
-const chaiHttp = require("chai-http");
-const expect = chai.expect;
-const should = chai.should();
-chai.use(chaiHttp);
-
-const constants = require('./constants');
-const expectUnchanged = require('./compare-responses').expectUnchanged;
-
-//
-//
-
-/**
- * Test that a hits search returns the same response as before.
- *
- * @param testName The name of the test (and file name of the expected response).
- * @param params The search parameters.
- */
-function expectHitsUnchanged(testName, params) {
-
-    // You can call this function with one string parameter, which is then used
-    // as both the name and the CQL pattern.
-    if (params === undefined)
-        throw 'Please pass both a test name and CQL pattern (or parameter object)';
-
-    // You can specify a CQL pattern or a map of parameters
-    if (typeof params === 'string')
-        params = { patt: params };
-
-    describe(testName, () => {
-        it('response should match previous', done => {
-            chai.request(constants.SERVER_URL)
-            .get('/test/hits')
-            .query({
-                sort: "field:pid,hitposition", // fully defined sort
-                wordsaroundhit: 1,
-                waitfortotal: "true",
-                //usecache: "no", // causes the search to be executed multiple times (hits, count, etc.)
-                ...params
-            })
-            .set('Accept', 'application/json')
-            .end((err, res) => {
-                expect(err).to.be.null;
-                expect(res).to.have.status(200);
-                expectUnchanged('hits', testName, res.body);
-                done();
-            });
-        });
-    });
-}
-
-
-// Single word
-expectHitsUnchanged("single word the", '"the"');
-expectHitsUnchanged("simple phrase a succesful", '"a" [lemma="successful"]');
-// Also test that forward index matching either the first or the second clause produces the same results
-expectHitsUnchanged("phrase a succesful with _FI1", '_FI1("a", [lemma="successful"])');
-expectHitsUnchanged("phrase a succesful with _FI2", '_FI2("a", [lemma="successful"])');
-
-// Simple capture group
-expectHitsUnchanged("simple capture group", '"one" A:[]');
-
-// A few simpler tests, just checking matching text
-expectHitsUnchanged("any token", '[]');
-expectHitsUnchanged("two-four-single-regex", '"two|four"');
-expectHitsUnchanged("two-four-separate", '"two"|"four"');
-expectHitsUnchanged("token level AND", '[lemma="be" & word="are"]');
-expectHitsUnchanged("token level AND NOT", '[lemma="be" & word!="are"]');
-expectHitsUnchanged("containing", '<u/> containing "good"');
-expectHitsUnchanged("within", '"very" "good" within <u/>');
-
-// View a single group from grouped hits
-expectHitsUnchanged('view single group', {
-    patt: '"a"',
-    group: 'field:title',
-    viewgroup: 'str:service encounter about visa application for family members',
-});
